LayoutTests/http/tests/security/xss-DENIED-window-name-navigator-expected.txt

CONSOLE MESSAGE: line 1: SecurityError (DOM Exception 18): Sandbox access violation: Blocked a frame at "null" from accessing a frame at "http://127.0.0.1:8000".  The frame requesting access is sandboxed and lacks the "allow-same-origin" flag.
